@@ -1,8 +1,8 @@
 package com.james602152002.floatinglabelspinner.adapter;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnitRunner;
+import androidx.test.platform.app.InstrumentationRegistry;
+import androidx.test.runner.AndroidJUnitRunner;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
