@@ -1,7 +1,7 @@
 package com.james602152002.floatinglabelspinner.popupwindow;
 
 import android.content.Context;
-import android.support.v7.widget.CardView;
+import androidx.cardview.widget.CardView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.AdapterView;
